added error messages

+++ src/main/java/edu/sjsu/cs249/project1/client/ClientApplication.java
@@ -25,7 +25,7 @@ public class ClientApplication {
         serverService.register(client1);
         System.out.println("+ Client Started +");
 
-        String command = "";
+        String command = "help";
         String help = "Please enter a command. Available commands:\n" +
                 "ls - Lists the available files\n" +
                 "create <filename> <contents>- create a file with the given name \n"+
@@ -41,7 +41,7 @@ public class ClientApplication {
             String input = scanner.nextLine();
             String[] inputs = input.split(" ");
             command = inputs[0].toLowerCase();
-
+            
             switch (command) {
                 case "help": {
                     System.out.println(help);
@@ -62,6 +62,10 @@ public class ClientApplication {
                 //  through server.
                 case "open": {
                     byte[] temp = null;
+                    if (inputs.length < 2) {
+                        System.out.println("Error. No file name given.");
+                        break;
+                    }
                     String fileName = inputs[1];
                     System.out.println("Opening: " + fileName);
                     File file = client1.getCachedFile(fileName);
@@ -76,7 +80,6 @@ public class ClientApplication {
                             client1.readFiles(temp);
                         }
                     }
-
                     break;
                 }
 
@@ -84,6 +87,10 @@ public class ClientApplication {
                 // cache will be removed as well.
 
                 case "rm": {
+                    if (inputs.length < 2) {
+                        System.out.println("Error. No file name given.");
+                        break;
+                    }
                     String fileName = inputs[1];
                     System.out.println("Deleting: " + fileName);
                     serverService.removeFiles(client1, fileName);
@@ -94,7 +101,15 @@ public class ClientApplication {
                 // create a file on the server and add to local cache.
 
                 case "create": {
+                    if (inputs.length < 2) {
+                        System.out.println("Error. No file name given.");
+                        break;
+                    }
                     String fileName = inputs[1];
+                    if (inputs.length < 3) {
+                        System.out.println("Error. No data given.");
+                        break;
+                    }
                     String data = input.substring(input.indexOf(inputs[2]));
                     byte[] contents = data.getBytes(Charset.forName("UTF-8"));
                     System.out.println("Creating: " + fileName);
@@ -105,7 +120,15 @@ public class ClientApplication {
 
                 // Replace the byte array and update local cache as well.
                 case "modify": {
+                    if (inputs.length < 2) {
+                        System.out.println("Error. No file name given.");
+                        break;
+                    }
                     String fileName = inputs[1];
+                    if (inputs.length < 3) {
+                        System.out.println("Error. No data given.");
+                        break;
+                    }
                     String data = input.substring(input.indexOf(inputs[2]));
                     byte[] newContents = data.getBytes(Charset.forName("UTF-8"));
                     System.out.println("Modifying: " + fileName);
@@ -116,7 +139,15 @@ public class ClientApplication {
 
                 // Replace the file name with a new file name
                 case "rename": {
+                    if (inputs.length < 2) {
+                        System.out.println("Error. No file name given.");
+                        break;
+                    }
                     String fileName = inputs[1];
+                    if (inputs.length < 3) {
+                        System.out.println("Error. No new file name given.");
+                        break;
+                    }
                     String newName = inputs[2];
                     System.out.println("Renaming: " + fileName + " to: " + newName);
                     serverService.renameFile(client1, fileName, newName);
@@ -131,8 +162,13 @@ public class ClientApplication {
                     break;
                 }
 
+                case "": {
+                    System.out.println("Enter an input.");
+                    break;
+                }
+
                 default:
-                    System.out.println("Invalid command. Enter 'help' for help.");
+                    System.out.println("Invalid command: [" + command + "]  Enter 'help' for help.");
             }
         }
